+++ crates/cli/tests/reference/constructor.d.ts
@@ -2,6 +2,6 @@
 /* eslint-disable */
 export class ClassConstructor {
   free(): void;
-  [symbolDispose](): void;
+  [Symbol.dispose](): void;
   constructor();
 }