+++ crates/cli/tests/reference/echo.js
@@ -1206,7 +1206,7 @@ const FooFinalization = (typeof FinalizationRegistry === 'undefined')
     ? { register: () => {}, unregister: () => {} }
     : new FinalizationRegistry(ptr => wasm.__wbg_foo_free(ptr >>> 0, 1));
 
-const symbolDispose = Symbol.dispose || Symbol('Symbol.dispose');
+const symbolDispose = Symbol.dispose || Symbol.for('wbg_symbol_dispose');
 
 export class Foo {