+++ crates/cli/tests/reference/raw.d.ts
@@ -4,7 +4,7 @@ export function test1(test: number): number;
 export class Test {
   private constructor();
   free(): void;
-  [symbolDispose](): void;
+  [Symbol.dispose](): void;
   static test1(test: number): Test;
   test2(test: number): void;
 }