+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 * Added bindings for `WeakRef`.
   [#4659](https://github.com/wasm-bindgen/wasm-bindgen/pull/4659)
+* Support `Symbol.dispose` by default, without overwriting the global `Symbol.dispose`
+  when unsupported, using `Symbol.for('wbg_symbol_dispose')` when unavailable.
 
 ### Fixed